LayoutTests/http/tests/security/denied-base-url-javascript-url-expected.txt

CONSOLE MESSAGE: Blocked setting javascript:// This is JavaScript as the base URL because it is not a valid URL.

PASS 'javascript:' is an invalid base URL.
